navbar: Remove padding-top from #tab_bar CSS because it defaults to 0.
Prior to commit 2a263a60b0c407ad9b5a1dbb84f644abdc71c010 the center
area of the navbar was based on a structure that appended crumbs or
"tabs" as <li>s, forming a tab_bar and a tab_list.

However, in 2a263a60b0c407ad9b5a1dbb84f644abdc71c010 we apply a new
style and structure to the navbar which lets go of the convention of
tabs. Hence, we'd like to purge the tab_bar and tab_list labels from
our code base. This commit pushes us towards that goal by simplifying
the future merge of the two CSS labels.