refactor: Remove redundant JQuery checking block.
We can safely remove the block added in commit
7d75a17c1be2be276cde82e58f4e4555d099a497
which checks whether the target is a JQuery object
and applies it, if it is not, because re-applying
the JQuery selector on a JQuery object returns the
same object.

We also refactor the related function calls to pass
the target instead of applying the JQuery selector
to it and finding the closest "subscription_settings"
class, as the same operation takes place in the next
line of the removed block. Thus this is redundant too.