node tests: Show coverage link when we regress coverage.
In 2a3e36609d98dec020d748f8def5d20ba08cc301 we added
the convenient line to tell folks about the coverage
report.  But if we failed coverage checks, we didn't
show the link.

Arguably we should just always show this, even if
tests fail, but that can also be potentially confusing.