emojisets: Prefetch the octopus via Webpack.
As of commit 37143c3a711a91db8073e737121d7982d156c9ce we use the
octopus image through Webpack, so the prefetch needs to be from the
same Webpack URL for it to do any good.  We don’t want to waste more
bandwidth on this [AWESOME CRITICAL FEATURE] than we have to.

Signed-off-by: Anders Kaseorg <anders@zulip.com>