minor: Remove confusing parens.
Apparently I put these parens in the code as
part of 30be576650dc6c67883b47c3bcef8b26f827e59c
during 2017.

It looks like I extracted is_public during
the middle of my change and forgot to remove
the unnecessary parens.  (The code was correct,
but it makes it look like a tuple if you're
skimming it too quickly.)