dev_panel: Use correct content helper for portico pages.
In 9906438d397af9aadaea45c87fe578a9ad27336a, this page was converted
to use portico styling, but we intentionally left this page not using
the portico_content class since we didn't want the header/footer.

We still don't want the header/footer clutter, so instead, we achieve
that same goal using the isolated_page flag.